fix: revert ProtocolResolver rename
