+++ lib/QueryRunner.js
@@ -1,5 +1,5 @@
 import { lookup } from './game-resolver.js'
-import { getProtocol } from './protocol-resolver.js'
+import { getProtocol } from './ProtocolResolver.js'
 import GlobalUdpSocket from './GlobalUdpSocket.js'
 
 const defaultOptions = {
